split provider dependencies
