+++ packages/mina-graphql/src/Providers/index.ts
@@ -1,6 +1,7 @@
 export * from './AccountInfo'
 export * from './ChainHistory'
 export * from './provider'
+export * from './providerArchive'
 export * from './TxStatus'
 export * from './TxSubmit'
 export * from './types'
