+++ packages/mina-graphql/src/Providers/types.ts
@@ -9,8 +9,10 @@ import {
   TxStatus,
   TxStatusArgs
 } from '@palladxyz/mina-core'
-
-export interface Provider {
+/**
+ * An interface that abstracts over Mina node queries.
+ */
+export interface ProviderNode {
   /**
    * The provider itself
    */
@@ -27,16 +29,33 @@ export interface Provider {
   // getTransactionCount(address: AddressLike, blockTag?: BlockTag): Promise<number>;
   // getBlock(blockHashOrBlockTag: BlockTag | string, prefetchTxs?: boolean): Promise<null | Block>;
   // getTransactionResult(hash: string): Promise<null | string>;
+  
+  getAccountInfo(args: AccountInfoArgs): Promise<AccountInfo> 
 
-  getAccountInfo(args: AccountInfoArgs): Promise<AccountInfo>
+  getTransactionStatus(args: TxStatusArgs): Promise<TxStatus>
+
+  submitTransaction(args: SubmitTxArgs): Promise<SubmitTxResult>
+}
+
+/**
+ * An interface that abstracts over Mina Archive node queries.
+ */
+export interface ProviderArchive {
+  /**
+   * The provider itself
+   */
+  provider: this
+
+  /**
+   *  Shutdown any resources this provider is using. No additional
+   *  calls should be made to this provider after calling this.
+   */
+  destroy(): void
 
   getTransactions(
     args: TransactionsByAddressesArgs
   ): Promise<Mina.Paginated<Mina.TransactionBody>>
 
-  getTransactionStatus(args: TxStatusArgs): Promise<TxStatus>
-
   getTransaction(args: TransactionsByIdsArgs): Promise<Mina.TransactionBody[]>
 
-  submitTransaction(args: SubmitTxArgs): Promise<SubmitTxResult>
 }
